ecs_worker/requirements.txt
boto3>=1.34.0
Pillow>=10.1.0